Cloud function/requirements.txt
google-api-python-client
google-auth
python-dateutil
Flask
functions-framework
Cloud function main.py
Cloud function
